#180 - Pdf program: checks the presence of a session before downloading its program + addition of an error log

+++ BatchOperations/Generic/MailingBatchOperation.php
@@ -106,37 +106,42 @@ class MailingBatchOperation extends AbstractBatchOperation implements BatchOpera
         $this->setOptions($options);
         $deleteTemplate = false;
 
-        //---setting choosed template file
-        // 1/ File was provided by user
-        if (isset($this->options['attachment']) && !empty($this->options['attachment'])) {
-            $attachment = $this->options['attachment'][0];
-            $attachment->move($this->options['tempDir'], $attachment->getClientOriginalName());
-            $this->currentTemplate = $this->options['tempDir'].$attachment->getClientOriginalName();
-            $this->currentTemplateFileName = $attachment->getClientOriginalName();
-            $deleteTemplate = true;
-        } elseif (isset($this->options['template']) && (is_integer($this->options['template']))) {
-            //file was choosed in template list
-            $templateTerm = $this->container->get('sygefor_core.vocabulary_registry')->getVocabularyById('sygefor_core.vocabulary_publipost_template');
-            /** @var EntityManager $em */
-            $em = $this->em;
-            $repo = $em->getRepository(get_class($templateTerm));
-            /** @var PublipostTemplate[] $templates */
-            $template = $repo->find($this->options['template']);
+        if (count($entities)) {
+            //---setting choosed template file
+            // 1/ File was provided by user
+            if (isset($this->options['attachment']) && !empty($this->options['attachment'])) {
+                $attachment = $this->options['attachment'][0];
+                $attachment->move($this->options['tempDir'], $attachment->getClientOriginalName());
+                $this->currentTemplate = $this->options['tempDir'].$attachment->getClientOriginalName();
+                $this->currentTemplateFileName = $attachment->getClientOriginalName();
+                $deleteTemplate = true;
+            } elseif (isset($this->options['template']) && (is_integer($this->options['template']))) {
+                //file was choosed in template list
+                $templateTerm = $this->container->get('sygefor_core.vocabulary_registry')->getVocabularyById('sygefor_core.vocabulary_publipost_template');
+                /** @var EntityManager $em */
+                $em = $this->em;
+                $repo = $em->getRepository(get_class($templateTerm));
+                /** @var PublipostTemplate[] $templates */
+                $template = $repo->find($this->options['template']);
 
-            $this->currentTemplate = $template->getAbsolutePath();
-            $this->currentTemplateFileName = $template->getFileName();
-        } else {
-            // 3/ Error...
-            return '';
+                $this->currentTemplate = $template->getAbsolutePath();
+                $this->currentTemplateFileName = $template->getFileName();
+            } else {
+                // 3/ Error...
+                return '';
+            }
+
+            $parseInfos = $this->parseFile($this->currentTemplate, $entities);
+
+            if ($deleteTemplate) {
+                unlink($this->currentTemplate);
+            }
+
+            return $parseInfos;
         }
-
-        $parseInfos = $this->parseFile($this->currentTemplate, $entities);
-
-        if ($deleteTemplate) {
-            unlink($this->currentTemplate);
-        }
-
-        return $parseInfos;
+        
+        $this->container->get('monolog.logger.batch_operation')->error("MailingBatchOperation[execute] - The idList parameter is an empty array but it shouldn't be.");
+        return;
     }
 
     /**
